Move concept to replace the original api
