+++ src/ewatercycle/base/model.py
@@ -1,98 +1,138 @@
+import abc
 import logging
-from abc import ABCMeta, abstractmethod
-from datetime import datetime
-from typing import Any, ClassVar, Generic, Iterable, Optional, Tuple, TypeVar
+from abc import abstractmethod
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import Any, Iterable, Type
 
+import bmipy
 import numpy as np
+import pydantic
 import xarray as xr
-from bmipy import Bmi
+import yaml
 from cftime import num2date
 from grpc4bmi.bmi_optionaldest import OptionalDestBmi
 from grpc4bmi.reserve import reserve_values, reserve_values_at_indices
 
-from ewatercycle._repr import Representation
 from ewatercycle.base.forcing import DefaultForcing
 from ewatercycle.base.parameter_set import ParameterSet
+from ewatercycle.config import CFG
+from ewatercycle.container import start_container
+from ewatercycle.util import to_absolute_path
 
 logger = logging.getLogger(__name__)
-
-ForcingT = TypeVar("ForcingT", bound=DefaultForcing)
 
 
 ISO_TIMEFMT = r"%Y-%m-%dT%H:%M:%SZ"
 
 
-class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
-    """Abstract class of a eWaterCycle model."""
+class BaseModel(pydantic.BaseModel, abc.ABC):
+    """Base functionality for eWaterCycle models.
 
-    available_versions: ClassVar[Tuple[str, ...]] = tuple()
-    """Versions of model that are available in this class"""
+    Children need to specify how to make their BMI instance: in a container or
+    local python environment.
+    """
 
-    def __init__(
-        self,
-        version: str,
-        parameter_set: Optional[ParameterSet] = None,
-        forcing: Optional[ForcingT] = None,
-    ):
-        self.version = version
-        self.parameter_set = parameter_set
-        self.forcing: Optional[ForcingT] = forcing
-        self._check_version()
+    forcing: DefaultForcing
+    parameter_set: ParameterSet
+    parameters: dict[str, Any]
+    _bmi: bmipy.Bmi = pydantic.PrivateAttr()
+
+    @abc.abstractmethod
+    def _make_bmi_instance(self) -> bmipy.Bmi:
+        """Attach a BMI instance to self._bmi"""
+
+    def __post_init_post_parse__(self):
+        """Check model compatibility."""
         self._check_parameter_set()
-        self.bmi: Bmi = None  # bmi should set in setup() before calling its methods
-        """Basic Modeling Interface object"""
+
+    def setup(self, *, cfg_dir: str | None = None, **kwargs) -> tuple[str, str]:
+        """Performs model setup.
+
+        1. Creates config file and config directory
+        2. Start bmi instance and store as self._bmi
+
+        Args:
+            cfg_dir: Optionally specify path to use as config dir. Will be
+                created if it doesn't exist yet. Behaviour follows PyMT
+                documentation
+                (https://pymt.readthedocs.io/en/latest/usage.html#model-setup).
+                Only difference is that we don't create a temporary directory,
+                but rather a time-stamped folder inside
+                ewatercycle.CFG['output_dir'].
+            *args: Positional arguments. Sub class should specify each arg.
+            **kwargs: Named arguments. Sub class should specify each arg.
+
+        Returns:
+            Path to config file and path to config directory
+        """
+        self.cfg_dir: Path = self._make_cfg_dir(cfg_dir)
+        self.cfg_file = self._make_cfg_file(**kwargs)
+        self._bmi = self._make_bmi_instance()
+
+        return str(self.cfg_file), str(self.cfg_dir)
+
+    def _make_cfg_dir(self, cfg_dir):
+        if cfg_dir is not None:
+            cfg_dir = to_absolute_path(cfg_dir)
+        else:
+            tz = timezone.utc
+            timestamp = datetime.now(tz).strftime("%Y%m%d_%H%M%S")
+            cfg_dir = to_absolute_path(
+                f"ewatercycle_{timestamp}", parent=CFG.output_dir
+            )
+
+        cfg_dir.mkdir(parents=True, exist_ok=True)
+
+        return cfg_dir
+
+    def _make_cfg_file(self, **kwargs):
+        cfg_file = self.cfg_dir / "config.yaml"
+        self.parameters.update(**kwargs)
+        with open(cfg_file, "w") as file:
+            yaml.dump({k: v for k, v in self.parameters}, file)
+
+        return cfg_file
 
     def __del__(self):
         try:
-            del self.bmi
+            del self._bmi
         except AttributeError:
             pass
 
     def __repr_args__(self):
         # Ignore bmi and internal state from subclasses
         return [
-            ("version", self.version),
             ("parameter_set", self.parameter_set),
             ("forcing", self.forcing),
         ]
 
-    @abstractmethod
-    def setup(self, *args, **kwargs) -> Tuple[str, str]:
-        """Performs model setup.
-        1. Creates config file and config directory
-        2. Start bmi container and store as self.bmi
-        Args:
-            *args: Positional arguments. Sub class should specify each arg.
-            **kwargs: Named arguments. Sub class should specify each arg.
-        Returns:
-            Path to config file and path to config directory
-        """
-
+    # BMI methods
     def initialize(self, config_file: str) -> None:
         """Initialize the model.
         Args:
             config_file: Name of initialization file.
         """
-        self.bmi.initialize(config_file)
+        self._bmi.initialize(config_file)
 
     def finalize(self) -> None:
         """Perform tear-down tasks for the model."""
-        self.bmi.finalize()
-        del self.bmi
+        self._bmi.finalize()
+        del self._bmi
 
     def update(self) -> None:
         """Advance model state by one time step."""
-        self.bmi.update()
+        self._bmi.update()
 
     def get_value(self, name: str) -> np.ndarray:
         """Get a copy of values of the given variable.
         Args:
             name: Name of variable
         """
-        if isinstance(self.bmi, OptionalDestBmi):
-            return self.bmi.get_value(name)
-        dest = reserve_values(self.bmi, name)
-        return self.bmi.get_value(name, dest)
+        if isinstance(self._bmi, OptionalDestBmi):
+            return self._bmi.get_value(name)
+        dest = reserve_values(self._bmi, name)
+        return self._bmi.get_value(name, dest)
 
     def get_value_at_coords(
         self, name, lat: Iterable[float], lon: Iterable[float]
@@ -105,10 +145,10 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
         """
         indices = self._coords_to_indices(name, lat, lon)
         indices = np.array(indices)
-        if isinstance(self.bmi, OptionalDestBmi):
-            return self.bmi.get_value_at_indices(name, indices)
-        dest = reserve_values_at_indices(self.bmi, name, indices)
-        return self.bmi.get_value_at_indices(name, dest, indices)
+        if isinstance(self._bmi, OptionalDestBmi):
+            return self._bmi.get_value_at_indices(name, indices)
+        dest = reserve_values_at_indices(self._bmi, name, indices)
+        return self._bmi.get_value_at_indices(name, dest, indices)
 
     def set_value(self, name: str, value: np.ndarray) -> None:
         """Specify a new value for a model variable.
@@ -116,7 +156,7 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
             name: Name of variable
             value: The new value for the specified variable.
         """
-        self.bmi.set_value(name, value)
+        self._bmi.set_value(name, value)
 
     def set_value_at_coords(
         self, name: str, lat: Iterable[float], lon: Iterable[float], values: np.ndarray
@@ -130,7 +170,7 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
         """
         indices = self._coords_to_indices(name, lat, lon)
         indices = np.array(indices)
-        self.bmi.set_value_at_indices(name, indices, values)
+        self._bmi.set_value_at_indices(name, indices, values)
 
     def _coords_to_indices(
         self, name: str, lat: Iterable[float], lon: Iterable[float]
@@ -145,7 +185,7 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
             "not implemented for this model."
         )
 
-    @abstractmethod
+    @abstractmethod  # TODO: provide default implementation
     def get_value_as_xarray(self, name: str) -> xr.DataArray:
         """Get a copy values of the given variable as xarray DataArray.
         The xarray object also contains coordinate information and additional
@@ -154,39 +194,39 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
         """
 
     @property
-    @abstractmethod
-    def parameters(self) -> Iterable[Tuple[str, Any]]:
-        """Default values for the setup() inputs."""
+    def bmi(self) -> bmipy.Bmi:
+        """Bmi class wrapped by the model."""
+        return self._bmi
 
     @property
     def start_time(self) -> float:
         """Start time of the model."""
-        return self.bmi.get_start_time()
+        return self._bmi.get_start_time()
 
     @property
     def end_time(self) -> float:
         """End time of the model."""
-        return self.bmi.get_end_time()
+        return self._bmi.get_end_time()
 
     @property
     def time(self) -> float:
         """Current time of the model."""
-        return self.bmi.get_current_time()
+        return self._bmi.get_current_time()
 
     @property
     def time_units(self) -> str:
         """Time units of the model. Formatted using UDUNITS standard from Unidata."""
-        return str(self.bmi.get_time_units())
+        return str(self._bmi.get_time_units())
 
     @property
     def time_step(self) -> float:
         """Current time step of the model."""
-        return self.bmi.get_time_step()
+        return self._bmi.get_time_step()
 
     @property
     def output_var_names(self) -> Iterable[str]:
         """List of a model's output variables."""
-        return self.bmi.get_output_var_names()
+        return self._bmi.get_output_var_names()
 
     @property
     def start_time_as_isostr(self) -> str:
@@ -213,8 +253,8 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
     def start_time_as_datetime(self) -> datetime:
         """Start time of the model as a datetime object."""
         return num2date(
-            self.bmi.get_start_time(),
-            self.bmi.get_time_units(),
+            self._bmi.get_start_time(),
+            self._bmi.get_time_units(),
             only_use_cftime_datetimes=False,
         )
 
@@ -222,8 +262,8 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
     def end_time_as_datetime(self) -> datetime:
         """End time of the model as a datetime object'."""
         return num2date(
-            self.bmi.get_end_time(),
-            self.bmi.get_time_units(),
+            self._bmi.get_end_time(),
+            self._bmi.get_time_units(),
             only_use_cftime_datetimes=False,
         )
 
@@ -231,8 +271,8 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
     def time_as_datetime(self) -> datetime:
         """Current time of the model as a datetime object'."""
         return num2date(
-            self.bmi.get_current_time(),
-            self.bmi.get_time_units(),
+            self._bmi.get_current_time(),
+            self._bmi.get_time_units(),
             only_use_cftime_datetimes=False,
         )
 
@@ -246,22 +286,39 @@ class AbstractModel(Generic[ForcingT], Representation, metaclass=ABCMeta):
                 f"Parameter set has wrong target model, "
                 f"expected {model_name} got {self.parameter_set.target_model}"
             )
-        if self.parameter_set.supported_model_versions == set():
-            logger.info(
-                f"Model version {self.version} is not explicitly listed in the "
-                "supported model versions of this parameter set. "
-                "This can lead to compatibility issues."
-            )
-        elif self.version not in self.parameter_set.supported_model_versions:
-            raise ValueError(
-                "Parameter set is not compatible with version {self.version} of "
-                "model, parameter set only supports "
-                f"{self.parameter_set.supported_model_versions}."
-            )
 
-    def _check_version(self):
-        if self.version not in self.available_versions:
-            raise ValueError(
-                f"Supplied version {self.version} is not supported by this model. "
-                f"Available versions are {self.available_versions}."
-            )
+
+class LocalModel(BaseModel):
+    """eWaterCycle model running in a local Python environment.
+
+    Mostly intended for development purposes.
+    """
+
+    bmi_class: Type[bmipy.Bmi]
+
+    def _make_bmi_instance(self):
+        return self.bmi_class()
+
+
+class ContainerizedModel(BaseModel):
+    """eWaterCycle model running inside a container.
+
+    This is the recommended method for sharing eWaterCycle models.
+    """
+
+    bmi_image: str = "ghcr.io/ewatercycle/leakybucket-grpc4bmi:latest"
+
+    def _make_bmi_instance(self) -> bmipy.Bmi:
+        self.additional_input_dirs = []
+        if self.parameter_set:
+            self.additional_input_dirs.append(str(self.parameter_set.directory))
+        if self.forcing:
+            self.additional_input_dirs.append(str(self.forcing.directory))
+
+        grpc4bmi = start_container(
+            image_engine=self.bmi_image,  # TODO: find way to infer image name based on CFG['container_engine']
+            work_dir=self.cfg_dir,
+            input_dirs=self.additional_input_dirs,
+            timeout=300,
+        )
+        return grpc4bmi
